Upload of the install and documentation
Documentation was written to meet the criteria of the coursework at the time in 2022. It covers the SDLC of the project with analysis, design, implementation, testing and evaluation. Plenty of diagrams show the entity relationships with regards to SQL, the classes and interfaces in the program, flow of algorithms, etc.